requirements_dev.txt
-r requirements.txt
ipython
pdbpp
django-extensions
django-pdb
graphviz
